+++ src/test/java/Autotests/duck_controller/UpdateDuck.java
@@ -1,0 +1,126 @@
+package Autotests.duck_controller;
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.annotations.CitrusResource;
+import com.consol.citrus.annotations.CitrusTest;
+import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.testng.annotations.Optional;
+import org.testng.annotations.Test;
+
+import static com.consol.citrus.DefaultTestActionBuilder.action;
+import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
+import static com.consol.citrus.http.actions.HttpActionBuilder.http;
+
+public class UpdateDuck extends TestNGCitrusSpringSupport{
+
+    @Test(description = "Изменение цвета и высоты уточки",enabled = true)
+    @CitrusTest
+    public void UpdateColorAndHeight(@Optional @CitrusResource TestCaseRunner runner)
+    {
+        createDuck(runner,"yellow","1.0","wood","quack","FIXED");
+        runner.$(http().client("http://localhost:2222")
+                .receive()
+                .response(HttpStatus.OK)
+                .message()
+                .extract(fromBody().expression("$.id", "duckId")));
+        updateDuck(runner,"${duckId}","red","2.0","wood","quack","FIXED");
+        validateResponse(runner,"{\n" + "  \"message\": \"Duck with id = ${duckId} is updated\"\n" + "}");
+        deleteDuck(runner,"${duckId}");
+    }
+
+    @Test(description = "Изменение цвета и звука уточки",enabled = true)
+    @CitrusTest
+    public void UpdateColorAndSound(@Optional @CitrusResource TestCaseRunner runner)
+    {
+        createDuck(runner,"yellow","1.0","wood","quack","FIXED");
+        runner.$(http().client("http://localhost:2222")
+                .receive()
+                .response(HttpStatus.OK)
+                .message()
+                .extract(fromBody().expression("$.id", "duckId")));
+        updateDuck(runner,"${duckId}","red","1.0","wood","roar","FIXED");
+        validateResponse(runner,"{\n" + "  \"message\": \"Duck with id = ${duckId} is updated\"\n" + "}");
+        deleteDuck(runner,"${duckId}");
+    }
+
+    public void duckSwim(TestCaseRunner runner, String id) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .get("/api/duck/action/swim")
+                .queryParam("id", id));
+    }
+
+    public void createDuck(TestCaseRunner runner, String color, String height, String material, String sound, String wingsState) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .post("/api/duck/create")
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body("{\n" + "  \"color\": \"" + color + "\",\n"
+                        + "  \"height\": " + height + ",\n"
+                        + "  \"material\": \"" + material + "\",\n"
+                        + "  \"sound\": \"" + sound + "\",\n"
+                        + "  \"wingsState\": \"" + wingsState
+                        + "\"\n" + "}"));
+    }
+
+    public void deleteDuck(TestCaseRunner runner, String id) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .delete("/api/duck/delete")
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .queryParam("id",id));
+    }
+    public void validateResponse(TestCaseRunner runner, String responseMessage) {
+        runner.$(http().client("http://localhost:2222")
+                .receive()
+                .response(HttpStatus.OK)
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(responseMessage));
+    }
+
+    public void updateDuck(TestCaseRunner runner,String id, String color, String height, String material, String sound, String wingsState) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .put("/api/duck/update")
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .queryParam("id",id)
+                .queryParam("color",color)
+                .queryParam("height",height)
+                .queryParam("material",material)
+                .queryParam("sound",sound)
+                .queryParam("wingsState",wingsState));
+
+    }
+
+    public void quackDuck(TestCaseRunner runner,String id, String repetitionCount, String soundCount) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .get("/api/duck/action/quack")
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .queryParam("id",id)
+                .queryParam("repetitionCount",repetitionCount)
+                .queryParam("soundCount",soundCount));
+
+    }
+
+    public void duckFly(TestCaseRunner runner, String id) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .get("/api/duck/action/fly")
+                .queryParam("id", id));
+    }
+
+    public void showProperties(TestCaseRunner runner, String id) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .get("/api/duck/action/properties")
+                .queryParam("id", id));
+    }
+
+}
